hdfsee: Boost, as given on command line, can be a non-integer

+++ src/dw-hdfsee.c
@@ -1778,7 +1778,7 @@ static gint displaywindow_press(GtkWidget *widget, GdkEventButton *event,
 
 
 DisplayWindow *displaywindow_open(const char *filename, const char *peaks,
-                                  int boost, int binning, int cmfilter,
+                                  double boost, int binning, int cmfilter,
                                   int noisefilter, int colscale,
                                   const char *element, const char *geometry,
                                   int show_rings, double *ring_radii,
